Add utility functions.
